fix: rename module files
